Implement findBlockSize

+++ src/lib/aes.hpp
@@ -4,7 +4,7 @@
 #include <span>
 #include <vector>
 
-constexpr int BLOCK_SIZE = 16;
+constexpr size_t BLOCK_SIZE = 16;
 constexpr int ROUNDS = 10;
 
 std::vector<uint8_t> padPKCS7(std::span<const uint8_t> buffer);
